implemented remove item from cart with ajax

+++ src/Controller/ShoppingCartController.php
@@ -9,6 +9,7 @@ use App\Repository\ShoppingCartItemRepository;
 use App\Repository\UserRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
@@ -28,23 +29,10 @@ class ShoppingCartController extends AbstractController
     }
     #[Route('/addToCart/{productId}', name: 'addToCart')]
 
-    public function addToCart(Request $request, UserRepository $userRepository, ProductRepository $productRepository, $productId): Response
+    public function addToCart(ProductRepository $productRepository, $productId): Response
     {
         $entityManager = $this->getDoctrine()->getManager();
 
-        // Get the current session
-        // $session = $request->getSession();
-        // if (!$session->isStarted()) {
-        //     $session->start();
-        // }
-
-        // // get the email of the authenticated user
-        // $email = $session->get('_security.last_username');
-        // if (!$email) {
-        //     return new Response('where email', Response::HTTP_BAD_REQUEST);
-        // }
-        // //find the user instance in the database
-        // $user = $userRepository->findOneBy(['email' => $email]);
         $user = $this->getUser();
         if (!$user) {
             return new Response('Utilisateur introuvable', Response::HTTP_BAD_REQUEST);
@@ -55,7 +43,6 @@ class ShoppingCartController extends AbstractController
         if (!$product) {
             return new Response("Ce produit n'existe pas", Response::HTTP_BAD_REQUEST);
         }
-        // Check if the user already has the product in their cart
         $existingCartItem = $entityManager->getRepository(ShoppingCartItem::class)
             ->findOneBy([
                 'user' => $user,
@@ -81,22 +68,10 @@ class ShoppingCartController extends AbstractController
     }
     #[Route('/cart', name: 'cart')]
 
-    public function cart(Request $request, UserRepository $userRepository, ShoppingCartItemRepository $shoppingCartRepository): Response
+    public function cart(ShoppingCartItemRepository $shoppingCartRepository): Response
     {
-        $entityManager = $this->getDoctrine()->getManager();
 
-        // Get the current session
-        // $session = $request->getSession();
-        // if (!$session->isStarted()) {
-        //     $session->start();
-        // }
 
-        // // get the email of the authenticated user
-        // $email = $session->get('_security.last_username');
-        // if (!$email) {
-        //     return new Response('where email', Response::HTTP_BAD_REQUEST);
-        // }
-        // $user = $userRepository->findOneBy(['email' => $email]);
         $user = $this->getUser();
 
         if (!$user) {
@@ -108,21 +83,23 @@ class ShoppingCartController extends AbstractController
     }
     #[Route('/removeItem/{id}', name: 'removeItem')]
 
-    public function removeItem(ManagerRegistry $doctrine, ShoppingCartItemRepository $shoppingCartRepository, $id,): Response
+    public function removeItem(ManagerRegistry $doctrine, ShoppingCartItemRepository $shoppingCartRepository, $id): JsonResponse
     {
         $user = $this->getUser();
         if (!$user) {
-            return new Response('Utilisateur introuvable', Response::HTTP_BAD_REQUEST);
+            return new JsonResponse(['success' => false, 'message' => 'Utilisateur introuvable'], JsonResponse::HTTP_BAD_REQUEST);
         }
         $product = $shoppingCartRepository->find($id);
+        if (!$product) {
+            return new JsonResponse(['success' => false, 'message' => 'Produit introuvable'], JsonResponse::HTTP_BAD_REQUEST);
+        }
         $em = $doctrine->getManager();
         $em->remove($product);
         $em->flush();
 
-
-
-        return $this->redirectToRoute('cart');
+        return new JsonResponse(['success' => true, 'message' => 'Product removed successfully']);
     }
+
     #[Route('/api/cart', name: 'cartApi')]
 
     public function cartApi(Request $request, UserRepository $userRepository, ShoppingCartItemRepository $shoppingCartRepository, NormalizerInterface $normalizer): Response
